flatten parser module
