+++ rust/candid_parser/build.rs
@@ -2,6 +2,6 @@ fn main() {
     lalrpop::Configuration::new()
         .use_cargo_dir_conventions()
         .emit_rerun_directives(true)
-        .process_file("src/parser/grammar.lalrpop")
+        .process_file("src/grammar.lalrpop")
         .unwrap();
 }
