+++ rust/candid_parser/src/grammar.rs
@@ -1,0 +1,2 @@
+#![allow(clippy::all)]
+include!(concat!(env!("OUT_DIR"), "/grammar.rs"));
