--- rust/candid_parser/src/parser/grammar.rs
@@ -1,2 +1,0 @@
-#![allow(clippy::all)]
-include!(concat!(env!("OUT_DIR"), "/parser/grammar.rs"));
