+++ rust/candid_parser/src/random.rs
