+++ rust/candid_parser/tests/parse_value.rs
@@ -1,7 +1,7 @@
 use candid::types::value::{IDLArgs, IDLField, IDLValue, VariantValue};
 use candid::types::{Label, Type, TypeEnv, TypeInner};
 use candid::{record, variant, CandidType, Nat};
-use candid_parser::parser::parse_idl_args;
+use candid_parser::parse_idl_args;
 
 fn parse_args(input: &str) -> IDLArgs {
     parse_idl_args(input).unwrap()
@@ -12,10 +12,10 @@ fn parse_args_err(input: &str) -> candid_parser::Result<IDLArgs> {
 }
 
 fn parse_type(input: &str) -> Type {
-    use candid_parser::parser::types::IDLType;
+    use candid_parser::types::IDLType;
     let env = TypeEnv::new();
     let ast = input.parse::<IDLType>().unwrap();
-    candid_parser::parser::typing::ast_to_type(&env, &ast).unwrap()
+    candid_parser::typing::ast_to_type(&env, &ast).unwrap()
 }
 
 #[test]
